requirements.txt
google-generativeai
google-ai-generativelanguage
google-adk
fastapi
uvicorn
python-dotenv
sse-starlette
PyGithub
